test(workflow): relocate guards + state-machine legacy suites to co-located siblings (#1705)
